+++ src/headless/plugins/bookmarks/tests/bookmarks.js
@@ -3,39 +3,191 @@ const { Strophe, sizzle, stx, u } = converse.env;
 
 describe("A chat room", function () {
 
-    describe("when autojoin is set", function () {
+    beforeEach(() => jasmine.addMatchers({ toEqualStanza: jasmine.toEqualStanza }));
 
-        it("will be be opened and joined automatically upon login", mock.initConverse(
-                [], {}, async function (_converse) {
+    it("is automatically bookmarked when opened", mock.initConverse(['chatBoxesFetched'], {}, async (_converse) => {
+        const { bare_jid } = _converse;
+        await mock.waitForRoster(_converse, 'current', 0);
+        await mock.waitUntilDiscoConfirmed(
+            _converse, bare_jid,
+            [{'category': 'pubsub', 'type': 'pep'}],
+            [
+                'http://jabber.org/protocol/pubsub#publish-options',
+                'urn:xmpp:bookmarks:1#compat'
+            ]
+        );
 
-            await mock.waitForRoster(_converse, 'current', 0);
-            await mock.waitUntilBookmarksReturned(_converse);
-            spyOn(_converse.api.rooms, 'create').and.callThrough();
-            const jid = 'theplay@conference.shakespeare.lit';
-            const { bookmarks } = _converse.state;
-            const model = bookmarks.create({
-                jid,
-                'autojoin': false,
-                'name':  'The Play',
-                'nick': ''
-            });
-            expect(_converse.api.rooms.create).not.toHaveBeenCalled();
-            bookmarks.remove(model);
-            bookmarks.create({
-                jid,
-                'autojoin': true,
-                'name':  'Hamlet',
-                'nick': ''
-            });
-            expect(_converse.api.rooms.create).toHaveBeenCalled();
-        }));
-    });
+        const nick = 'JC';
+        const muc_jid = 'theplay@conference.shakespeare.lit';
+        const settings = { name: "Play's the thing", password: 'secret' };
+        const muc = await mock.openAndEnterChatRoom(_converse, muc_jid, nick, [], [], true, settings);
+
+        const IQ_stanzas = _converse.api.connection.get().IQ_stanzas;
+        const sent_stanza = await u.waitUntil(
+            () => IQ_stanzas.filter(s => sizzle('iq publish[node="urn:xmpp:bookmarks:1"]', s).length).pop());
+
+        expect(sent_stanza).toEqualStanza(
+            stx`<iq from="${_converse.bare_jid}"
+                    to="${_converse.bare_jid}"
+                    id="${sent_stanza.getAttribute('id')}"
+                    type="set"
+                    xmlns="jabber:client">
+                <pubsub xmlns="http://jabber.org/protocol/pubsub">
+                    <publish node="urn:xmpp:bookmarks:1">
+                        <item id="${muc.get('jid')}">
+                            <conference xmlns="urn:xmpp:bookmarks:1" autojoin="true" name="${settings.name}">
+                                <nick>${nick}</nick>
+                                <password>${settings.password}</password>
+                            </conference>
+                        </item>
+                    </publish>
+                    <publish-options>
+                        <x type="submit" xmlns="jabber:x:data">
+                            <field type="hidden" var="FORM_TYPE">
+                                <value>http://jabber.org/protocol/pubsub#publish-options</value>
+                            </field>
+                            <field var='pubsub#persist_items'>
+                                <value>true</value>
+                            </field>
+                            <field var='pubsub#max_items'>
+                                <value>max</value>
+                            </field>
+                            <field var='pubsub#send_last_published_item'>
+                                <value>never</value>
+                            </field>
+                            <field var='pubsub#access_model'>
+                                <value>whitelist</value>
+                            </field>
+                        </x>
+                    </publish-options>
+                </pubsub>
+            </iq>`
+        );
+
+        /* Server acknowledges successful storage
+         * <iq to='juliet@capulet.lit/balcony' type='result' id='pip1'/>
+         */
+        const stanza = stx`<iq
+            xmlns="jabber:client"
+            to="${_converse.api.connection.get().jid}"
+            type="result"
+            id="${sent_stanza.getAttribute('id')}"/>`;
+        _converse.api.connection.get()._dataRecv(mock.createRequest(stanza));
+
+        expect(muc.get('bookmarked')).toBeTruthy();
+    }));
 });
 
 
 describe("A bookmark", function () {
 
     beforeEach(() => jasmine.addMatchers({ toEqualStanza: jasmine.toEqualStanza }));
+
+    describe("when autojoin is set", function () {
+
+        it("will cause a MUC to be opened and joined automatically upon login", mock.initConverse(
+                [], {}, async function (_converse) {
+
+            const { api } = _converse;
+            await mock.waitForRoster(_converse, 'current', 0);
+            await mock.waitUntilBookmarksReturned(_converse);
+            spyOn(_converse.api.rooms, 'create').and.callThrough();
+            const { bookmarks } = _converse.state;
+
+            let jid = 'theplay@conference.shakespeare.lit';
+            const model = bookmarks.create({
+                jid,
+                autojoin: false,
+                name:  'The Play',
+                nick: ''
+            });
+            expect(_converse.api.rooms.create).not.toHaveBeenCalled();
+
+            // Check that we don't auto-join if muc_respect_autojoin is false
+            api.settings.set('muc_respect_autojoin', false);
+            bookmarks.create({
+                jid,
+                autojoin: true,
+                name:  'The Play',
+                nick: ''
+            });
+            expect(_converse.api.rooms.create).not.toHaveBeenCalled();
+
+            api.settings.set('muc_respect_autojoin', true);
+            bookmarks.remove(model);
+            bookmarks.create({
+                jid,
+                autojoin: true,
+                name:  'Hamlet',
+                nick: ''
+            });
+            expect(_converse.api.rooms.create).toHaveBeenCalled();
+
+        }));
+
+        it("has autojoin set to false upon leaving", mock.initConverse([], {}, async function (_converse) {
+            const { u } = converse.env;
+            await mock.waitForRoster(_converse, 'current', 0);
+            await mock.waitUntilBookmarksReturned(_converse);
+
+            const nick = 'romeo';
+            const muc_jid = 'theplay@conference.shakespeare.lit';
+            const settings = { name:  'The Play' };
+            const muc = await mock.openAndEnterChatRoom(_converse, muc_jid, nick, [], [], true, settings);
+
+            const { bookmarks } = _converse.state;
+            await u.waitUntil(() => bookmarks.length);
+            await u.waitUntil(() => muc.get('bookmarked'));
+            spyOn(bookmarks, 'sendBookmarkStanza').and.callThrough();
+
+            const sent_IQs = _converse.api.connection.get().IQ_stanzas;
+            while (sent_IQs.length) { sent_IQs.pop(); }
+
+            await muc.close();
+            await u.waitUntil(() => sent_IQs.length);
+
+            // Check that an IQ stanza is sent out, containing no
+            // conferences to bookmark (since we removed the one and
+            // only bookmark).
+            const sent_stanza = sent_IQs.pop();
+            expect(sent_stanza).toEqualStanza(
+                stx`<iq from="${_converse.bare_jid}"
+                        to="${_converse.bare_jid}"
+                        id="${sent_stanza.getAttribute('id')}"
+                        type="set"
+                        xmlns="jabber:client">
+                    <pubsub xmlns="http://jabber.org/protocol/pubsub">
+                        <publish node="urn:xmpp:bookmarks:1">
+                            <item id="${muc_jid}">
+                                <conference xmlns="urn:xmpp:bookmarks:1" name="${settings.name}" autojoin="false">
+                                    <nick>${nick}</nick>
+                                </conference>
+                            </item>
+                        </publish>
+                        <publish-options>
+                            <x type="submit" xmlns="jabber:x:data">
+                                <field type="hidden" var="FORM_TYPE">
+                                    <value>http://jabber.org/protocol/pubsub#publish-options</value>
+                                </field>
+                                <field var='pubsub#persist_items'>
+                                    <value>true</value>
+                                </field>
+                                <field var='pubsub#max_items'>
+                                    <value>max</value>
+                                </field>
+                                <field var='pubsub#send_last_published_item'>
+                                    <value>never</value>
+                                </field>
+                                <field var='pubsub#access_model'>
+                                    <value>whitelist</value>
+                                </field>
+                            </x>
+                        </publish-options>
+                    </pubsub>
+                </iq>`
+            );
+        }));
+    });
 
     it("can be created and sends out a stanza", mock.initConverse(
             ['connected', 'chatBoxesFetched'], {}, async function (_converse) {
@@ -47,7 +199,7 @@ describe("A bookmark", function () {
         const muc1_jid = 'theplay@conference.shakespeare.lit';
         const { bookmarks } = _converse.state;
 
-        bookmarks.createBookmark({
+        bookmarks.setBookmark({
             jid: muc1_jid,
             autojoin: true,
             name:  'Hamlet',
@@ -90,7 +242,7 @@ describe("A bookmark", function () {
 
 
         const muc2_jid = 'balcony@conference.shakespeare.lit';
-        bookmarks.createBookmark({
+        bookmarks.setBookmark({
             jid: muc2_jid,
             autojoin: true,
             name:  'Balcony',
@@ -136,7 +288,7 @@ describe("A bookmark", function () {
             </iq>`);
 
         const muc3_jid = 'garden@conference.shakespeare.lit';
-        bookmarks.createBookmark({
+        bookmarks.setBookmark({
             jid: muc3_jid,
             autojoin: false,
             name:  'Garden',