+++ src/plugins/bookmark-views/components/bookmark-form.js
@@ -38,7 +38,7 @@ class MUCBookmarkForm extends CustomElement {
         ev.preventDefault();
         const { bookmarks } = _converse.state;
         const form = /** @type {HTMLFormElement} */ (ev.target);
-        bookmarks.createBookmark({
+        bookmarks.setBookmark({
             jid: this.jid,
             autojoin: /** @type {HTMLInputElement} */ (form.querySelector('input[name="autojoin"]'))?.checked || false,
             name: /** @type {HTMLInputElement} */ (form.querySelector('input[name=name]'))?.value,