+++ src/types/plugins/bookmark-views/mixins.d.ts
@@ -1,4 +1,4 @@
-export namespace bookmarkableChatRoomView {
+export namespace BookmarkableChatRoomView {
     /**
      * Set whether the groupchat is bookmarked or not.
      * @private